Add files via upload
The code generates random bits of 0's and 1's which will be used to generate symbols. The symbols generated here are  -1 and 1 for random bits 0 and 1 respectively. An arbitrary Gaussian noise is added at the transmitter side, mimicking the noise generated at the Transmitter ( Here the Gaussian noise added due to the transmitter has a standard deviation of 0.1). 
A simple channel is designed, where loss is swept from 1dB to 30 dB. The channel also adds a Gaussian noise of standard deviation=1. Later, direct detection is used to detect the signals sent with a detection efficiency of 0.8. 
The bit error rate with respect to the loss has been plotted. The bit error rate saturates at 50% as expected.